Factored apply out into accessor
