+++ holoviews/core/options.py
@@ -43,9 +43,8 @@ import numpy as np
 
 import param
 from .tree import AttrTree
-from .util import sanitize_identifier, group_sanitizer,label_sanitizer, basestring, OrderedDict, config
-from .util import deprecated_opts_signature, disable_constant
-from .pprint import InfoPrinter, PrettyPrinter
+from .util import sanitize_identifier, group_sanitizer,label_sanitizer, basestring, OrderedDict
+from .pprint import InfoPrinter
 
 
 def cleanup_custom_options(id, weakref=None):
@@ -90,160 +89,6 @@ class SkipRendering(Exception):
         self.warn = warn
         super(SkipRendering, self).__init__(message)
 
-
-class Opts(object):
-
-    def __init__(self, obj, mode=None):
-        self._mode = mode
-        self._obj = obj
-
-
-    def get(self, group=None, backend=None):
-        """Returns the corresponding Options object.
-
-        Args:
-            group: The options group. Flattens across groups if None.
-            backend: Current backend if None otherwise chosen backend.
-
-        Returns:
-            Options object associated with the object containing the
-            applied option keywords.
-        """
-        keywords = {}
-        groups = Options._option_groups if group is None else [group]
-        backend = backend if backend else Store.current_backend
-        for group in groups:
-            optsobj = Store.lookup_options(backend, self._obj, group)
-            keywords = dict(keywords, **optsobj.kwargs)
-        return Options(**keywords)
-
-
-    def __call__(self, *args, **kwargs):
-        """Applies nested options definition.
-
-        Applies options on an object or nested group of objects in a
-        flat format. Unlike the .options method, .opts modifies the
-        options in place by default. If the options are to be set
-        directly on the object a simple format may be used, e.g.:
-
-            obj.opts(cmap='viridis', show_title=False)
-
-        If the object is nested the options must be qualified using
-        a type[.group][.label] specification, e.g.:
-
-            obj.opts('Image', cmap='viridis', show_title=False)
-
-        or using:
-
-            obj.opts({'Image': dict(cmap='viridis', show_title=False)})
-
-        Args:
-            *args: Sets of options to apply to object
-                Supports a number of formats including lists of Options
-                objects, a type[.group][.label] followed by a set of
-                keyword options to apply and a dictionary indexed by
-                type[.group][.label] specs.
-            backend (optional): Backend to apply options to
-                Defaults to current selected backend
-            clone (bool, optional): Whether to clone object
-                Options can be applied in place with clone=False
-            **kwargs: Keywords of options
-                Set of options to apply to the object
-
-        For backwards compatibility, this method also supports the
-        option group semantics now offered by the hv.opts.apply_groups
-        utility. This usage will be deprecated and for more
-        information see the apply_options_type docstring.
-
-        Returns:
-            Returns the object or a clone with the options applied
-        """
-        if self._mode is None:
-            apply_groups, _, _ = deprecated_opts_signature(args, kwargs)
-            if apply_groups and config.future_deprecations:
-                msg = ("Calling the .opts method with options broken down by options "
-                       "group (i.e. separate plot, style and norm groups) is deprecated. "
-                       "Use the .options method converting to the simplified format "
-                       "instead or use hv.opts.apply_groups for backward compatibility.")
-                param.main.warning(msg)
-
-        return self._dispatch_opts( *args, **kwargs)
-
-    def _dispatch_opts(self, *args, **kwargs):
-        if self._mode is None:
-            return self._base_opts(*args, **kwargs)
-        elif self._mode == 'holomap':
-            return self._holomap_opts(*args, **kwargs)
-        elif self._mode == 'dynamicmap':
-            return self._dynamicmap_opts(*args, **kwargs)
-
-    def clear(self, clone=False):
-        """Clears any options applied to the object.
-
-        Args:
-            clone: Whether to return a cleared clone or clear inplace
-
-        Returns:
-            The object cleared of any options applied to it
-        """
-        return self._obj.opts(clone=clone)
-
-    def info(self, show_defaults=False):
-        """Prints a repr of the object including any applied options.
-
-        Args:
-            show_defaults: Whether to include default options
-        """
-        pprinter = PrettyPrinter(show_options=True, show_defaults=show_defaults)
-        print(pprinter.pprint(self._obj))
-
-    def _holomap_opts(self, *args, **kwargs):
-        clone = kwargs.pop('clone', None)
-        apply_groups, _, _ = deprecated_opts_signature(args, kwargs)
-        data = OrderedDict([(k, v.opts(*args, **kwargs))
-                             for k, v in self._obj.data.items()])
-
-        # By default do not clone in .opts method
-        if (apply_groups if clone is None else clone):
-            return self._obj.clone(data)
-        else:
-            self._obj.data = data
-            return self._obj
-
-    def _dynamicmap_opts(self, *args, **kwargs):
-        from ..util import Dynamic
-
-        clone = kwargs.get('clone', None)
-        apply_groups, _, _ = deprecated_opts_signature(args, kwargs)
-        # By default do not clone in .opts method
-        clone = (apply_groups if clone is None else clone)
-
-        obj = self._obj if clone else self._obj.clone()
-        dmap = Dynamic(obj, operation=lambda obj, **dynkwargs: obj.opts(*args, **kwargs),
-                       streams=self._obj.streams, link_inputs=True)
-        if not clone:
-            with disable_constant(self._obj):
-                obj.callback = self._obj.callback
-                self._obj.callback = dmap.callback
-            dmap = self._obj
-            dmap.data = OrderedDict([(k, v.opts(*args, **kwargs))
-                                     for k, v in self._obj.data.items()])
-        return dmap
-
-
-    def _base_opts(self, *args, **kwargs):
-        apply_groups, options, new_kwargs = deprecated_opts_signature(args, kwargs)
-
-        # By default do not clone in .opts method
-        clone = kwargs.get('clone', None)
-        if apply_groups:
-            from ..util import opts
-            if options is not None:
-                kwargs['options'] = options
-            return opts.apply_groups(self._obj, **dict(kwargs, **new_kwargs))
-
-        kwargs['clone'] = False if clone is None else clone
-        return self._obj.options(*args, **kwargs)
 
 
 class OptionError(Exception):
